Move to inline format arguments
This will be a Clippy warning in future.

Also fix a few other Clippy future warnings:
- Use Box::default()
- Drop unneeded casts
- Drop unneeded &

+++ hybrid/src/ecies_aead_hkdf_dem_helper.rs
@@ -112,7 +112,7 @@ impl subtle::EciesAeadHkdfDemHelper for EciesAeadHkdfDemHelper {
         &self,
         symmetric_key_value: &[u8],
     ) -> Result<tink_core::Primitive, tink_core::TinkError> {
-        if symmetric_key_value.len() != self.get_symmetric_key_size() as usize {
+        if symmetric_key_value.len() != self.get_symmetric_key_size() {
             return Err("symmetric key has incorrect length".into());
         }
         let mut sk = Vec::new();
